feat: add models for Practicum and ScreenButton in Django migration
The migration file adds models for Practicum and ScreenButton in the Django
application. This allows for the creation of Practicum instances with fields
such as title, image, description, priority, and speciality. Additionally, the
ScreenButton model includes fields for screen_id, button_title, side, screen
number, and URL. These models enhance the functionality of the application by
providing structured data storage for practicums and screen buttons.